% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance}
\alias{deviance}
\title{Calculate the Deviance

Calculates the deviance from the likelihood.}
\usage{
deviance(mod)
}
\arguments{
\item{mod}{is the lmer model object. For exmaple, summary(mod).}
}
\description{
Calculate the Deviance

Calculates the deviance from the likelihood.
}
