% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intraclass Correlation

Calculate the Intaclass Corrlation (iCC).}
\usage{
icc(mod)
}
\arguments{
\item{mod}{is the lmer model object. For exmaple, summary(mod).}
}
\description{
Intraclass Correlation

Calculate the Intaclass Corrlation (iCC).
}
