% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq1.R
\name{rsq1}
\alias{rsq1}
\title{R-Squared for Level-1

Calculates the proportion of level-1 vairance explained by the model.}
\usage{
rsq1(base, mod)
}
\arguments{
\item{base}{is the simpler lmer model object.}

\item{mod}{is the more complex lmer model object.}
}
\description{
R-Squared for Level-1

Calculates the proportion of level-1 vairance explained by the model.
}
