% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq2.R
\name{rsq2}
\alias{rsq2}
\title{R-Squared for Level-2

Calculates the proportion of level-2 vairance explained by the model.}
\usage{
rsq2(base, mod)
}
\arguments{
\item{base}{is the simpler lmer model object.}

\item{mod}{is the more complex lmer model object.}
}
\description{
R-Squared for Level-2

Calculates the proportion of level-2 vairance explained by the model.
}
