% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqSlo.R
\name{rsqSlo}
\alias{rsqSlo}
\title{R-Squared for Slope Variance

Calculates the proportion of slope vairance explained by the model.}
\usage{
rsqSlo(base, mod, par = 2)
}
\arguments{
\item{base}{is the simpler lmer model object.}

\item{mod}{is the more complex lmer model object.}

\item{par}{is the parameter number for the slope variance you want to test.}
}
\description{
R-Squared for Slope Variance

Calculates the proportion of slope vairance explained by the model.
}
