% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testvar.R
\name{testvar}
\alias{testvar}
\title{Significance Test for Random Effects

Calculates the p-value for the test of the slope variance and intercept-slope covariance using the likelihood ratio test method.}
\usage{
testvar(outp, p = "intercept", testcov = FALSE)
}
\arguments{
\item{outp}{is the lmer model object.}

\item{p}{is string of the variance component you want tested (place in double quotes).}

\item{testcov}{is FALSE by default for the variance component. If testcov = TRUE, the function retuns the test of the intercept-component covariance.}
}
\description{
Significance Test for Random Effects

Calculates the p-value for the test of the slope variance and intercept-slope covariance using the likelihood ratio test method.
}
